src/pastekit/proxy/plugins/requirements.txt
# Charles Plugin Python Dependencies
websocket-client>=1.6.0
